chore: test optimization
